requirements_dpo.txt
hydra-core>=1.2.0,<1.3
omegaconf>=2.2,<2.3
pyyaml==6.0.1
torchmetrics>=0.4.1rc0,<=0.10.3
trl==0.10.1
transformers==4.48.0
wandb
webdataset>=0.1.48,<=0.1.62
pandas
sentencepiece
youtokentome
h5py
ijson
matplotlib>=3.3.2
sacremoses
sacrebleu
einops
faiss-cpu
sentence_transformers>=2.3
nltk>=3.6.5
jieba
ftfy
gdown
inflect
jieba
opencc==1.1.6
pangu
rapidfuzz
pybind11
pytorch-lightning==2.5.0
ipadic
mecab-python3
protobuf==3.20.3
datasets
dill==0.3.8
nemo_toolkit==2.1.0
regex
requests<2.32.0
python-daemon
huggingface_hub>=0.27.1
multiprocess==0.70.16
numba<=0.60.0
numpy>=1.24.3,<=1.25.2
rouge_score
setuptools>=70.0
lightning==2.5.0
ml-dtypes==0.2.0